+++ packages/libs/lambda-at-edge/src/default-handler.ts
@@ -59,9 +59,10 @@ import { CloudFrontService } from "./services/cloudfront.service";
 import { S3Service } from "./services/s3.service";
 import { RevalidateHandler } from "./handler/revalidate.handler";
 import { RenderService } from "./services/render.service";
-import { debug } from "./lib/console";
+import { debug, isDevMode } from "./lib/console";
 
 process.env.PRERENDER = "true";
+process.env.DEBUGMODE = Manifest.enableDebugMode;
 
 interface FoundFallbackInterface {
   routeRegex: string;
@@ -215,7 +216,9 @@ const REVALIDATION_CONFIG = Object.values<RouteConfig>(
   PrerenderManifest.routes
 ).find((r) => typeof r.initialRevalidateSeconds === "number");
 
-const REVALIDATE_IN = REVALIDATION_CONFIG?.initialRevalidateSeconds || 4;
+const REVALIDATE_IN = isDevMode()
+  ? 1
+  : REVALIDATION_CONFIG?.initialRevalidateSeconds || 4;
 
 const REVALIDATIONS: RevalidationInterface = {};
 
@@ -397,7 +400,7 @@ export const handler = async (
   }
 
   if (isOriginResponse(event)) {
-    debug(`[origin-response] event: ${JSON.stringify(event)}`);
+    debug(`[handle-origin-response] event: ${JSON.stringify(event)}`);
     response = await handleOriginResponse({
       event,
       manifest,
@@ -405,7 +408,7 @@ export const handler = async (
       context
     });
   } else {
-    debug(`[origin-request] event: ${JSON.stringify(event)}`);
+    debug(`[handle-origin-request] event: ${JSON.stringify(event)}`);
     response = await handleOriginRequest({
       event,
       manifest,
@@ -434,6 +437,53 @@ export const handler = async (
   debug(`[origin] final response: ${JSON.stringify(response)}`);
 
   return response;
+};
+
+/**
+ * check if this url and query params need to rewrite. And rewrite it if get configuration form serverless.yml
+ * Now, we can only support 1 url params, like rewrite /index.html?page=[number] to /page/[number].html
+ * We can use querystring lib if we want to support more functions.
+ *
+ * For example,
+ *     urlRewrites:
+ *        - name: paginationRewrite
+ *          originUrl: /index.html?page=[number]
+ *          rewriteUrl: /page/[number].html
+ *
+ * @param manifest
+ * @param request
+ */
+const checkAndRewriteUrl = (
+  manifest: OriginRequestDefaultHandlerManifest,
+  request: CloudFrontRequest
+): void => {
+  debug(`[checkAndRewriteUrl] manifest: ${JSON.stringify(manifest)}`);
+  const rewrites = manifest.urlRewrites;
+  debug(`[checkAndRewriteUrl] rewriteList: ${JSON.stringify(rewrites)}`);
+
+  if (!rewrites || rewrites.length === 0) return;
+
+  debug(`[checkAndRewriteUrl] Before: ${request.uri}, ${request.querystring}`);
+
+  const requestParamName = request.querystring.split("=")[0];
+  const requestParamValue = request.querystring.split("=")[1];
+  const requestUri = request.uri.split(".")[0];
+  if (!requestParamName || !requestParamValue || !requestUri) return;
+
+  debug(
+    `[checkAndRewriteUrl] requestParamName: ${requestParamName}, requestParamValue: ${requestParamValue}，requestUri: ${requestUri}`
+  );
+  rewrites.forEach(({ originUrl, rewriteUrl }) => {
+    debug(
+      `[originUrl: ${originUrl}, rewriteUrl: ${rewriteUrl}, prefix: ${requestUri}?${requestParamName}= ]`
+    );
+    if (originUrl.startsWith(`${requestUri}?${requestParamName}=`)) {
+      request.uri = `${rewriteUrl.split("[")[0]}${requestParamValue}.html`;
+      request.querystring = "";
+    }
+  });
+
+  debug(`[checkAndRewriteUrl] After: ${request.uri}, ${request.querystring}`);
 };
 
 const handleOriginRequest = async ({
@@ -603,6 +653,7 @@ const handleOriginRequest = async ({
       s3Origin.path = `${basePath}/static-pages/${manifest.buildId}`;
       const pageName = uri === "/" ? "/index" : uri;
       request.uri = `${pageName}.html`;
+      checkAndRewriteUrl(manifest, request);
       debug(`[origin-request] is html of fallback, uri: ${request.uri}`);
     } else if (isDataReq) {
       // We need to check whether data request is unmatched i.e routed to 404.html or _error.js