+++ packages/serverless-components/nextjs-component/src/component.ts
@@ -243,7 +243,9 @@ class NextjsComponent extends Component {
           handler: inputs.handler
             ? `${inputs.handler.split(".")[0]}.js`
             : undefined,
-          authentication: inputs.authentication ?? undefined
+          authentication: inputs.authentication ?? undefined,
+          urlRewrites: inputs.urlRewrites ?? undefined,
+          enableDebugMode: inputs.enableDebugMode ?? false
         },
         nextStaticPath
       );