+++ packages/serverless-components/nextjs-component/types.d.ts
@@ -52,6 +52,8 @@ export type ServerlessComponentInputs = {
   authentication?: { username: string; password: string };
   imageOptimizer?: boolean;
   certificateArn?: string;
+  urlRewrites?: UrlRewriteOptions;
+  enableDebugMode?: boolean;
 };
 
 type CloudfrontOptions = Record<string, any>;
@@ -83,3 +85,5 @@ export type LambdaInput = {
   runtime: string;
   name?: string;
 };
+
+type UrlRewriteOptions = { originUrl: string; rewriteUrl: string }[];