+++ packages/serverless-components/nextjs-component/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.20.39](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@1.20.38...@getjerry/serverless-next@1.20.39) (2021-08-23)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+## [1.20.38](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@1.20.37...@getjerry/serverless-next@1.20.38) (2021-08-23)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+## [1.20.37](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@1.20.36...@getjerry/serverless-next@1.20.37) (2021-08-23)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+## [1.20.36](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@1.20.35...@getjerry/serverless-next@1.20.36) (2021-08-23)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+## [1.20.35](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@1.20.34...@getjerry/serverless-next@1.20.35) (2021-08-23)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+## [1.20.34](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@1.20.33...@getjerry/serverless-next@1.20.34) (2021-08-20)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+## [1.20.33](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@1.20.32...@getjerry/serverless-next@1.20.33) (2021-08-20)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+## [1.20.32](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@1.20.31...@getjerry/serverless-next@1.20.32) (2021-08-20)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+## [1.20.31](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@1.20.30...@getjerry/serverless-next@1.20.31) (2021-08-05)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+## [1.20.30](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@1.20.29...@getjerry/serverless-next@1.20.30) (2021-08-04)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+## [1.20.29](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@1.20.28...@getjerry/serverless-next@1.20.29) (2021-08-03)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+## [1.20.28](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@1.20.27...@getjerry/serverless-next@1.20.28) (2021-08-03)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+## [1.20.27](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@1.20.26...@getjerry/serverless-next@1.20.27) (2021-08-03)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+## [1.20.26](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@1.20.25...@getjerry/serverless-next@1.20.26) (2021-08-03)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
+## [1.20.25](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@1.20.24...@getjerry/serverless-next@1.20.25) (2021-08-03)
+
+**Note:** Version bump only for package @getjerry/serverless-next
+
 ## [1.20.24](https://github.com/getjerry/serverless-next.js/compare/@getjerry/serverless-next@1.20.23...@getjerry/serverless-next@1.20.24) (2021-07-02)
 
 **Note:** Version bump only for package @getjerry/serverless-next